+++ packages/core/src/components/text-input/text-input.css.ts
@@ -16,13 +16,13 @@ export const root = componentRecipe({
             outline: 0,
             borderRadius: vars.size.borderRadius['300'],
             boxShadow: `inset 0 0 0 0.0625rem ${boxShadowColor}`,
-            backgroundColor: vars.color.background['canvas-100'],
+            backgroundColor: vars.color.background['canvas-base'],
             paddingBlock: vars.size.space['000'],
 
-            color: vars.color.foreground['normal-200'],
+            color: vars.color.foreground['normal'],
 
             selectors: {
-                '&::placeholder': { color: vars.color.foreground['hint-100'] },
+                '&::placeholder': { color: vars.color.foreground['hint'] },
                 [when.invalid()]: { vars: { [boxShadowColor]: vars.color.border.danger } },
                 [when.readonly()]: { backgroundColor: vars.color.gray['200'] },
                 [when.disabled()]: { opacity: 0.32, pointerEvents: 'none' },