+++ packages/core/src/components/toast/toast.css.ts
@@ -2,7 +2,6 @@ import { createGlobalVar, fallbackVar } from '@vanilla-extract/css';
 import { calc } from '@vanilla-extract/css-utils';
 import type { RecipeVariants } from '@vanilla-extract/recipes';
 
-import { foregrounds } from '~/styles/mixins/foreground.css';
 import { componentRecipe, componentStyle } from '~/styles/mixins/layer-style.css';
 import { typography } from '~/styles/mixins/typography.css';
 import { vars } from '~/styles/themes.css';
@@ -124,13 +123,16 @@ export const root = componentRecipe({
     variants: {
         colorPalette: {
             danger: {
-                backgroundColor: vars.color.background['danger-200'],
+                backgroundColor: vars.color.background['danger'],
+                color: vars.color.foreground.staticWhite,
             },
             success: {
-                backgroundColor: vars.color.background['success-200'],
+                backgroundColor: vars.color.background['success'],
+                color: vars.color.foreground.staticWhite,
             },
             info: {
-                backgroundColor: vars.color.background['contrast-200'],
+                backgroundColor: vars.color.background['canvas-inverse'],
+                color: vars.color.foreground.inverse,
             },
         },
     },
@@ -156,14 +158,8 @@ export const content = componentStyle({
     },
 });
 
-export const title = componentStyle([
-    foregrounds({ color: 'white' }),
-    typography({ style: 'subtitle1' }),
-]);
+export const title = componentStyle([{ color: 'inherit' }, typography({ style: 'subtitle1' })]);
 
-export const description = componentStyle([
-    foregrounds({ color: 'white' }),
-    typography({ style: 'body2' }),
-]);
+export const description = componentStyle([{ color: 'inherit' }, typography({ style: 'body2' })]);
 
 export type RootVariants = NonNullable<RecipeVariants<typeof root>>;