+++ packages/core/vitest.config.ts
@@ -3,7 +3,7 @@ import path from 'node:path';
 import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
-    plugins: [vanillaExtractPlugin()],
+    plugins: [vanillaExtractPlugin() as never],
     resolve: {
         alias: {
             '~': path.resolve(__dirname, 'src'),