+++ packages/icons/src/components/symbol-icons/AustriaColorIcon/AustriaColorIcon.tsx
@@ -2,27 +2,25 @@ import IconBase from '~/components/icon-base';
 import type { IconProps } from '~/components/icon-base';
 
 const AustriaColorIcon = (props: IconProps) => (
-    <IconBase viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+    <IconBase xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 16 16" {...props}>
         <mask
-            id="mask0_9214_1300"
-            style={{ maskType: 'luminance' }}
+            id="vapor-icons-color-AustriaColorIcon__a"
+            width={16}
+            height={16}
+            x={0}
+            y={0}
             maskUnits="userSpaceOnUse"
-            x="0"
-            y="0"
-            width="16"
-            height="16"
+            style={{
+                maskType: 'luminance',
+            }}
         >
-            <path
-                d="M8 16C12.4183 16 16 12.4183 16 8C16 3.58172 12.4183 0 8 0C3.58172 0 0 3.58172 0 8C0 12.4183 3.58172 16 8 16Z"
-                fill="white"
-            />
+            <path fill="#fff" d="M8 16A8 8 0 1 0 8 0a8 8 0 0 0 0 16" />
         </mask>
-        <g mask="url(#mask0_9214_1300)">
-            <path d="M19.8579 0H-4.14209V5.33273H19.8579V0Z" fill="#C8102E" />
-            <path d="M19.8579 5.3327H-4.14209V10.6654H19.8579V5.3327Z" fill="white" />
-            <path d="M19.8579 10.6673H-4.14209V16H19.8579V10.6673Z" fill="#C8102E" />
+        <g mask="url(#vapor-icons-color-AustriaColorIcon__a)">
+            <path fill="#C8102E" d="M19.858 0h-24v5.333h24z" />
+            <path fill="#fff" d="M19.858 5.333h-24v5.332h24z" />
+            <path fill="#C8102E" d="M19.858 10.667h-24V16h24z" />
         </g>
     </IconBase>
 );
-
 export default AustriaColorIcon;