+++ scripts/sync-figma/src/icons/templates/icon/icons-index.ts
@@ -1,0 +1,3 @@
+export default (iconNames: string[]) => `
+${iconNames.map((iconName) => `export { default as ${iconName} } from './${iconName}';`).join(' ')}
+`;